pyproject.toml
[tool.poetry]
name = "rag-app"
version = "0.1.0"
description = ""
authors = ["derkvandendoel <derk.vandendoel@collow.nl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.8"
streamlit = "^1.42.0"
langchain = "^0.3.18"
python-dotenv = "^1.0.1"
openai = "^1.64.0"
pymupdf4llm = "^0.0.17"
chromadb = "^0.6.3"
watchdog = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
